+++ client/src/dolbyio_rest_apis/media/analyze.py
@@ -26,12 +26,6 @@ async def start(
 
     See: https://docs.dolby.io/media-apis/reference/media-analyze-post
 
-    Beta API
-
-    This API is being made available as an early preview.
-    If you have feedback on how you'd like to use the API please reach out to share your feedback with our team.
-    https://dolby.io/contact
-
     Content Length
 
     Media content with duration less than 2 seconds will not be processed. The API will return an ERROR in this case.
@@ -48,11 +42,6 @@ async def start(
         HttpRequestError: If a client error one occurred.
         HTTPError: If one occurred.
     """
-    print('''Beta API
-    This API is being made available as an early preview.
-    If you have feedback on how you\'d like to use the API please reach out to share your feedback with our team.
-    https://dolby.io/contact''')
-
     async with MediaHttpContext() as http_context:
         json_response = await http_context.requests_post(
             access_token=access_token,
