+++ client/src/dolbyio_rest_apis/media/models/access_token.py
@@ -1,8 +1,8 @@
 """
-dolbyio_rest_apis.communications.models
+dolbyio_rest_apis.models
 ~~~~~~~~~~~~~~~
 
-This module contains the models used by the Dolby.io APIs.
+This module contains the models used by the Dolby.io MEDIA APIs.
 """
 
 from dolbyio_rest_apis.core.helpers import get_value_or_default
@@ -17,4 +17,3 @@ class AccessToken(dict):
         self.access_token = get_value_or_default(self, 'access_token', None)
         self.refresh_token = get_value_or_default(self, 'refresh_token', None)
         self.expires_in_val = get_value_or_default(self, 'expires_in', 0)
-        self.scope = get_value_or_default(self, 'scope', None)
